requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.12.2
    # via -r requirements.in
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via pycmarkgfm
charset-normalizer==2.0.12
    # via requests
idna==3.4
    # via requests
jinja2==3.1.1
    # via -r requirements.in
markupsafe==2.1.2
    # via jinja2
pycmarkgfm==1.2.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
python-dotenv==0.20.0
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
soupsieve==2.4.1
    # via beautifulsoup4
urllib3==1.26.9
    # via requests
